refactor(plugin): extract vault-crypto.ts (pure compute, scanner-clean)
Phase 1, Task 1.1 of the OpenClaw native integration
(docs/plans/2026-06-21-openclaw-native-integration-plan.md).

Physically separates pure-compute crypto into vault-crypto.ts so that
NO single source file co-contains environment-variable reads and
outbound-network primitives — the OpenClaw runtime scanner's
env-harvesting rule. vault-crypto.ts takes all key material and nonces
as parameters and reads the environment NOWHERE and performs network
I/O NOWHERE.

- vault-crypto.ts (NEW): consolidates the pure primitives previously
  split across crypto.ts (XChaCha20-Poly1305 via WASM: deriveKeys,
  encrypt/decrypt, generateBlindIndices, generateContentFingerprint,
  deriveLshSeed, computeAuthKeyHash, isBip39Mnemonic) and pair-crypto.ts
  (x25519 ECDH + HKDF-SHA256 + AES-256-GCM via node:crypto:
  generateGatewayKeypair, deriveSessionKeys, deriveAeadKeyFromEcdh,
  aeadEncryptWithSessionKey/aeadDecrypt, encryptPairingPayload/
  decryptPairingPayload, compareSecondaryCodesCT).
- crypto.ts, pair-crypto.ts: now thin re-export shims so the 331KB
  index.ts and other importers (tr-cli, pair-cli, pair-http,
  pair-remote-client, tr-cli-export-helper, pair-cli-relay) are not
  broken in this pass. Behaviour-preserving — bodies moved verbatim.
- vault-crypto.test.ts (NEW): asserts the hard contracts —
  (1) round-trip decrypt(encrypt(x,key),key)===x for both cipher suites,
  (2) generateBlindIndices returns 32-byte SHA-256 hex digests,
  (3) source-contract: vault-crypto.ts contains neither the env-var
      token nor the network-primitive token (regex mirrors the scanner
      rule, built from fragments so the test file itself stays clean).

ERC-4337 signing stays in subgraph-store.ts for now: the WASM
signUserOp call is pure compute but is deeply intertwined with the
bundler sponsor/send RPC flow. Extracting just that one-liner is a
no-op refactor that belongs with the relay/network re-org (Task 1.2).
Noted as follow-up in vault-crypto.ts module header.

Verification:
  npx tsx vault-crypto.test.ts  -> 8 checks passed (+1 SKIP)
  npm run check-scanner        -> 145 files, 0 flags
  npm run build                -> succeeds (re-exports resolve)
